+++ frontend/src/pages/Employees/EmployeesPage.js
@@ -39,8 +39,8 @@ class EmployeesPage extends Component {
 
     handleButtonClick(employee) {
         if (employee && employee.idEmployee) {
-            this.props.navigate(`/employee-details/`, {
-                state: { employee },
+            this.props.navigate("/employee-details", {
+                state: { idEmployee: employee.idEmployee },
             });
         } else {
             console.error("Données de l'employé non définies");